# capstone_project/docs/swagger.yaml
definitions:
  category.GameCategoryCreateDTO:
    properties:
      imageUrl:
        type: string
      name:
        type: string
    required:
    - imageUrl
    - name
    type: object
  category.GameCategoryResponseDTO:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  dto.LoginDTO:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  dto.LoginResponseDTO:
    properties:
      avatarUrl:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  dto.RegisterDTO:
    properties:
      avatarUrl:
        type: string
      description:
        type: string
      email:
        type: string
      gender:
        enum:
        - MALE
        - FEMALE
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - avatarUrl
    - description
    - email
    - gender
    - password
    - username
    type: object
  dto.RegisterResponseDTO:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
...
  game.GameRequestCreateDTO:
    properties:
      gameId:
        example: 1
        type: integer
    required:
    - gameId
    type: object
...
  game.GameRequestResponseDTO:
    properties:
      gameId:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      status:
        example: pending
        type: string
      userId:
        example: 1
        type: integer
    type: object
  game.GameRequestUpdateDTO:
    properties:
      status:
        example: accepted
        type: string
    required:
    - status
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with username/email and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponseDTO'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User registration information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponseDTO'
      summary: Create a new user
      tags:
      - auth
  /game-categories:
    get:
      consumes:
      - application/json
      description: Get a list of all game categories with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/category.GameCategoryResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all game categories with pagination
      tags:
      - game-categories
    post:
      consumes:
      - application/json
      description: Create a new game category with the provided details
      parameters:
      - description: Game category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/category.GameCategoryCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.GameCategoryResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new game category
      tags:
      - game-categories
  /game-requests:
    post:
      consumes:
      - application/json
      description: Create a new game request with status pending
      parameters:
      - default: Bearer <Add access token here>
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game request details
        in: body
        name: gameRequest
        required: true
        schema:
          $ref: '#/definitions/game.GameRequestCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/game.GameRequestResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new game request
      tags:
      - game-requests
  /game-requests/{id}:
    put:
      consumes:
      - application/json
      description: Approve or reject a game request
      parameters:
      - default: Bearer <Add access token here>
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Game request details to update
        in: body
        name: gameRequest
        required: true
        schema:
          $ref: '#/definitions/game.GameRequestUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.GameRequestResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a game request
      tags:
      - game-requests
...
swagger: "2.0"